feat: Add leaderboard screen and contest result type
Introduces a new `leaderboard` view to the application, allowing users to see contest results. This also includes the addition of a `ContestResult` type to represent individual participant outcomes within a contest. The Gemini service schemas have been updated to be dynamically generated based on the number of questions. Icon sizes and relevant types have been adjusted.